Reactive Interaction Gateway v2.4.0 RIG.DistributedSet

Distributed grow-only set with per key time-to-live support.

Types

creation_ts()

creation_ts() :: non_neg_integer()

expiration_ts()

expiration_ts() :: non_neg_integer()

key()

key() :: String.t()

record()

record() :: {key(), uuid(), creation_ts(), expiration_ts()}

uuid()

uuid() :: String.t()

Functions

add(server, key, ttl_s \\ 86400)

add(pid() | atom(), key(), non_neg_integer()) :: :ok

Add a new record to this set.

Re-adding already included keys may be used to extend a key's lifetime.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

handle_call(msg, from, state)

Handles adding a record on this node.

handle_info(msg, state)

Remove expired records.

has?(name, key, opts \\ [])

has?(pid() | atom(), key(), list()) :: boolean()

Check whether this set includes a key or not.

Returns true if, and only if, the key has been added previously and it has not yet expired.

start_link(set_name, opts \\ [])

Creates a new distributed set.

Parameters

set_name: The name of the distributed set. Multiple instances with the same name will synchronize their records with each other.