RIG.Subscriptions (Reactive Interaction Gateway v3.0.0-alpha.2)

Event subscriptions.

Types

claims()

Specs

claims() :: %{optional(String.t()) => String.t()}

jwt_conf()

Specs

jwt_conf() :: %{alg: String.t(), key: String.t()}

Functions

config()

Specs

config() :: any()

Returns module configuration.

If application environment contains values in Keyword or Map struct, defaults from macro usage will be recursively merged with application configuration.

If one of the configuration parameters does not exist or can not be resolved, raises ArgumentError.

from_json(json)

Specs

from_json(json :: String.t() | nil) ::
  Result.t([Rig.Subscription.t()], %RIG.Subscriptions.Error{
    __exception__: term(),
    cause: term()
  })

from_jwt_claims(claims, extractor_path_or_json \\ config().extractor_path_or_json)

from_token(token, jwt_conf \\ config().jwt_conf)

Specs

from_token(token :: RIG.JWT.token(), jwt_conf()) ::
  Result.t([Rig.Subscription.t()], %RIG.Subscriptions.Error{
    __exception__: term(),
    cause: term()
  })