Result (Reactive Interaction Gateway v3.0.0-alpha.1)

Tools for working with result tuples.

Influenced by the Rust Option/Result implementation.

Types

err()

Specs

err() :: err(any())

err(error_type)

Specs

err(error_type) :: {:error, error_type}

error()

Specs

error() :: any()

fmt_errors()

Specs

fmt_errors() :: ([error()] -> any())

ok()

Specs

ok() :: ok(any())

ok(value_type)

Specs

ok(value_type) :: {:ok, value_type}

t()

Specs

t() :: t(any(), any())

t(value_type, error_type)

Specs

t(value_type, error_type) :: ok(value_type) | err(error_type)

value()

Specs

value() :: any()

Functions

and_then(result, value_to_result_fn)

Specs

and_then(t(), (value() -> t())) :: t()

Maps a Result to another Result by applying a function to a contained value, leaving non-ok tuples untouched.

Note that the given function is expected to return a result. See Result.map/2 if you want to pass a function that returns a value.

Examples

iex> {:ok, :a} |> Result.and_then(fn :a -> {:ok, :b} end)
{:ok, :b}
iex> {:ok, :a} |> Result.and_then(fn :a -> {:error, :b} end)
{:error, :b}
iex> {:error, :a} |> Result.and_then(fn :a -> {:ok, :b} end)
{:error, :a}

err(error_or_err_result)

Specs

err(error() | err()) :: err()

Turns an error into a result that contains that error.

If the error is already a result, it is returned unchanged.

Examples

iex> Result.err(:a)
{:error, :a}
iex> Result.err(Result.err(:a))
{:error, :a}

err?(result)

Specs

err?(t()) :: boolean()

Returns false if a result has a contained value and true otherwise.

Examples

iex> Enum.filter([ok: 1, error: 2, ok: 3], &Result.err?/1)
[error: 2]

filter_and_unwrap(results)

Specs

filter_and_unwrap([t()]) :: [value()]

Unwraps ok-results and rejects error-results.

Examples

iex> [ok: :a, ok: :b, error: :c] |> Result.filter_and_unwrap()
[:a, :b]

filter_and_unwrap_err(results)

Specs

filter_and_unwrap_err([t()]) :: [error()]

Unwraps error-results and rejects ok-results.

Examples

iex> [ok: :a, ok: :b, error: :c] |> Result.filter_and_unwrap_err()
[:c]

list_to_result(results, fmt_errors \\ &(&1))

Specs

list_to_result([t()], fmt_errors()) :: ok(list()) | err(list())

Turns a list of results into a result with lists of either values or errors.

As soon as there is at least one error in the given list, the result is an error-type Result.

Example

iex> [ok: :a, ok: :b] |> Result.list_to_result()
{:ok, [:a, :b]}
iex> [ok: :a, ok: :b, error: :c, error: :d] |> Result.list_to_result()
{:error, [:c, :d]}

map(result, value_to_value_fn)

Specs

map(t(), (value() -> value())) :: t()

Maps a Result to another Result by applying a function to a contained value, leaving non-ok tuples untouched.

Note that the given function is expected to return a value. See Result.and_then/2 if you want to pass a function that returns a result.

Examples

iex> {:ok, :a} |> Result.map(fn :a -> :b end)
{:ok, :b}
iex> {:error, :a} |> Result.map(fn :a -> :b end)
{:error, :a}

map_err(result, error_to_error_fn)

Specs

map_err(t(), (error() -> error())) :: t()

Maps a Result to another Result by applying a function to a contained error, leaving ok tuples untouched.

This function can be used to compose the results of two functions, where the map function returns an error.

Examples

iex> {:error, :a} |> Result.map_err(fn :a -> :b end)
{:error, :b}
iex> {:ok, :a} |> Result.map_err(fn :a -> :b end)
{:ok, :a}

ok(value_or_ok_result)

Specs

ok(value() | ok()) :: ok()

Turns a value into a result that contains that value.

If the value is already a result, it is returned unchanged.

Examples

iex> Result.ok(:a)
{:ok, :a}
iex> Result.ok(Result.ok(:a))
{:ok, :a}

ok?(result)

Specs

ok?(t()) :: boolean()

Returns true if a result has a contained value and false otherwise.

Examples

iex> Enum.filter([ok: 1, error: 2, ok: 3], &Result.ok?/1)
[ok: 1, ok: 3]

iex> Enum.split_with([ok: 1, error: 2, ok: 3], &Result.ok?/1)
{[ok: 1, ok: 3], [error: 2]}

or_else(result, error_to_result_fn)

Specs

or_else(t(), (error() -> t())) :: t()

Maps a Result to another Result by applying a function to a contained value, leaving ok tuples untouched.

Note that the given function is expected to return a result. See Result.map_err/2 if you want to pass a function that returns a value.

Examples

iex> {:ok, :a} |> Result.or_else(fn :a -> {:ok, :b} end)
{:ok, :a}
iex> {:error, :a} |> Result.or_else(fn :a -> {:ok, :b} end)
{:ok, :b}
iex> {:error, :a} |> Result.or_else(fn :a -> {:error, :b} end)
{:error, :b}

result_to_list(arg)

Specs

result_to_list(t()) :: [ok()] | [err()]

Turns a result that holds a list of values or errors into a list of results.

Example

iex> {:ok, [:a, :b]} |> Result.result_to_list()
[ok: :a, ok: :b]
iex> {:error, [:c, :d]} |> Result.result_to_list()
[error: :c, error: :d]

unwrap(ok_result, expectation \\ nil)

Specs

unwrap(ok(), expectation :: String.t() | nil) :: value()

Returns the contained value or throw an error.

unwrap_err(err_result)

Specs

unwrap_err(err()) :: error()

Returns the contained error or throw an error.

unwrap_err_or(result, default_value)

Specs

unwrap_err_or(t(), default :: value() | error()) :: value() | error()

Returns the contained error or a default.

unwrap_or(result, default_value)

Specs

unwrap_or(t(), default :: value()) :: value()

Returns the contained value or a default.