Reactive Interaction Gateway v2.1.1 Rig.Connection.Codec

Encode and decode a connection token, e.g., for correlation.

Functions

config()

config() :: any()

Returns module configuration.

If application environment contains values in Keyword or Map struct, defaults from macro usage will be recursively merged with application configuration.

If one of the configuration parameters does not exist or can not be resolved, raises ArgumentError.

decrypt(ciphertext, key)

decrypt(binary(), binary()) :: binary()

Decrypts ciphertext using key. Ciphertext is init. vector, ciphertag (MAC) and the actual ciphertext concatenated.

deserialize(base64_encoded)

deserialize(binary()) :: {:ok, pid()} | {:error, :not_base64 | :invalid_term}

Convert a serialized string back into a pid.

deserialize!(bin)

deserialize!(binary()) :: pid()

Convert a serialized string back into a pid.

encrypt(val, key)

encrypt(binary(), binary()) :: binary()

Encrypts value using key and returns init. vector, ciphertag (MAC) and ciphertext concatenated. Additional authenticated data (AAD) adds parameters like protocol version num. to MAC

hash(key)

hash(binary()) :: binary()

The function processes any string or number and returns exactly 16 byte binary. The hash produced by this function can be used as key by both encrypt and decrypt functions.

serialize(pid)

serialize(pid()) :: binary()

Turn a pid into an url-encoded string.