Reactive Interaction Gateway v2.0.2 Rig.Discovery

Module responsible to start Peerage application if relevant. Peerage will be auto-configured by values from environment variables.

Functions

config()

config() :: any()

Returns module configuration.

If application environment contains values in Keyword or Map struct, defaults from macro usage will be recursively merged with application configuration.

If one of the configuration parameters does not exist or can not be resolved, raises ArgumentError.

start()