Reactive Interaction Gateway v2.1.0 Rig.EventFilter.Server

Filter for the event subscription mechanism.

For each event type, on each node, there is exactly one EventFilter (Server). The Filter keeps track of subscriptions. Subscriptions have a time-to-live, that is, if they're not renewed for some time, the Filter will remove them eventually.

Types

event_type()

event_type() :: String.t()

Functions

add_wildcards_to_table(table, n_cur_fields, n_new_fields)

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

process(event_type)

remove_expired_records(map)

start(event_type, config, opts \\ [])

start(event_type(), Rig.EventFilter.Config.event_type_config(), list()) ::
  {:ok, pid()}

start_link(event_type, config, opts \\ [])

start_link(event_type(), Rig.EventFilter.Config.event_type_config(), list()) ::
  {:ok, pid()}