Reactive Interaction Gateway v2.2.1 Rig.EventFilter.Sup

Starts filter processes on the local node on demand.

Each node should run exactly one EventFilter Supervisor.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

config()

config() :: any()

Returns module configuration.

If application environment contains values in Keyword or Map struct, defaults from macro usage will be recursively merged with application configuration.

If one of the configuration parameters does not exist or can not be resolved, raises ArgumentError.

processes()

start_link(opts \\ [])