Reactive Interaction Gateway v2.0.2 RigApi

The entrypoint for defining your web interface, such as controllers, views, channels and so on.

This can be used in your application as:

use RigApi, :controller
use RigApi, :view

The definitions below will be executed for every view, controller, etc, so keep them short and clean, focused on imports, uses and aliases.

Do NOT define functions inside the quoted expressions below. Instead, define any helper function in modules and import those modules here.

Functions

__using__(which) (macro)

When used, dispatch to the appropriate controller/view/etc.

controller()

router()

view()