RigApi.Endpoint (Reactive Interaction Gateway v3.0.0-alpha.1)

Functions

broadcast(topic, event, msg)

Callback implementation for Phoenix.Endpoint.broadcast/3.

broadcast!(topic, event, msg)

Callback implementation for Phoenix.Endpoint.broadcast!/3.

broadcast_from(from, topic, event, msg)

Callback implementation for Phoenix.Endpoint.broadcast_from/4.

broadcast_from!(from, topic, event, msg)

Callback implementation for Phoenix.Endpoint.broadcast_from!/4.

call(conn, opts)

Callback implementation for Plug.call/2.

child_spec(opts)

Returns the child specification to start the endpoint under a supervision tree.

config()

Specs

config() :: any()

Returns module configuration.

If application environment contains values in Keyword or Map struct, defaults from macro usage will be recursively merged with application configuration.

If one of the configuration parameters does not exist or can not be resolved, raises ArgumentError.

config(key, default \\ nil)

Returns the endpoint configuration for key

Returns default if the key does not exist.

config_change(changed, removed)

Reloads the configuration given the application environment changes.

host()

Returns the host for the given endpoint.

init(opts)

Callback implementation for Plug.init/1.

local_broadcast(topic, event, msg)

Callback implementation for Phoenix.Endpoint.local_broadcast/3.

local_broadcast_from(from, topic, event, msg)

Callback implementation for Phoenix.Endpoint.local_broadcast_from/4.

path(path)

Generates the path information when routing to this endpoint.

script_name()

Generates the script name.

start_link(opts \\ [])

Starts the endpoint supervision tree.

static_integrity(path)

Generates a base64-encoded cryptographic hash (sha512) to a static file in priv/static. Meant to be used for Subresource Integrity with CDNs.

static_lookup(path)

Returns a two item tuple with the first item being the static_path and the second item being the static_integrity.

static_path(path)

Generates a route to a static file in priv/static.

static_url()

Generates the static URL without any path information.

It uses the configuration under :static_url to generate such. It falls back to :url if :static_url is not set.

struct_url()

Generates the endpoint base URL but as a URI struct.

It uses the configuration under :url to generate such. Useful for manipulating the URL data and passing it to URL helpers.

subscribe(topic, opts \\ [])

unsubscribe(topic)

url()

Generates the endpoint base URL without any path information.

It uses the configuration under :url to generate such.