Reactive Interaction Gateway v2.2.1 RigApi.Router.Helpers

Module with named helpers generated from RigApi.Router.

Functions

apis_path(conn_or_endpoint, action)

apis_path(conn_or_endpoint, action, params)

apis_path(conn_or_endpoint, action, id, params)

apis_url(conn_or_endpoint, action)

apis_url(conn_or_endpoint, action, params)

apis_url(conn_or_endpoint, action, id, params)

channels_path(conn_or_endpoint, action)

channels_path(conn_or_endpoint, action, params)

channels_path(conn_or_endpoint, action, user, params)

channels_url(conn_or_endpoint, action)

channels_url(conn_or_endpoint, action, params)

channels_url(conn_or_endpoint, action, user, params)

health_path(conn_or_endpoint, action)

health_path(conn_or_endpoint, action, params)

health_url(conn_or_endpoint, action)

health_url(conn_or_endpoint, action, params)

message_path(conn_or_endpoint, action)

message_path(conn_or_endpoint, action, params)

message_url(conn_or_endpoint, action)

message_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

responses_path(conn_or_endpoint, action)

responses_path(conn_or_endpoint, action, params)

responses_url(conn_or_endpoint, action)

responses_url(conn_or_endpoint, action, params)

session_blacklist_path(conn_or_endpoint, action)

session_blacklist_path(conn_or_endpoint, action, params)

session_blacklist_path(conn_or_endpoint, action, session_id, params)

session_blacklist_url(conn_or_endpoint, action)

session_blacklist_url(conn_or_endpoint, action, params)

session_blacklist_url(conn_or_endpoint, action, session_id, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.