Reactive Interaction Gateway v2.1.0 RigApi.SessionBlacklistController

Allows for blocking "sessions" for a specific period of time.

What a session is depends on your business context and the JWT_SESSION_FIELD setting. For example, a session ID could be a random ID assigned to a token upon login, or the id of the user the token belongs to.

Functions

blacklist_session(conn, params)

Plug action to add a session id to the session blacklist.

check_status(conn, map)

Check blacklist status for a specific session id.

swagger_definitions()

swagger_path_blacklist_session(route)

swagger_path_check_status(route)