Reactive Interaction Gateway v3.0.0-alpha.1 RigApi.V3.APIs

CRUD controller for the reverse-proxy settings.

Functions

add_api(conn, params)

config()

config() :: any()

Returns module configuration.

If application environment contains values in Keyword or Map struct, defaults from macro usage will be recursively merged with application configuration.

If one of the configuration parameters does not exist or can not be resolved, raises ArgumentError.

deactivate_api(conn, params)

get_api_detail(conn, params)

list_apis(conn, params)

swagger_definitions()

swagger_path_add_api(route)

swagger_path_deactivate_api(route)

swagger_path_get_api_detail(route)

swagger_path_list_apis(route)

swagger_path_update_api(route)

update_api(conn, params)