Reactive Interaction Gateway v2.1.1 RigAuth.Blacklist

Enables blacklisting of JWTs by their jti claim.

The entries representing the banned claims feature an expiration timestamp, which prevents the blacklist from growing indefinitely.

In a distributed setting, the node that does the blacklisting spreads the information via Phoenix' PubSub Server as Phoenix Presence information. The other nodes react by tracking the same record themselves, which means that for one record and n nodes there are n items in the Presence list. The following properties are a result of this:

Blacklisting can occur on/by any node.

The blacklist is eventually consistent over all nodes.

Any node can go down and come up at any time without affecting the blacklist, except if all nodes go down at the same time (in that case there is nothing to synchronize from -- the list is not stored on disk).

Functions

add_jti(server, jti, expiry \\ nil, listener \\ nil)

add_jti(
  pid() | atom(),
  String.t(),
  nil | String.t() | Timex.DateTime.t(),
  nil | pid()
) :: pid()

config()

config() :: any()

Returns module configuration.

If application environment contains values in Keyword or Map struct, defaults from macro usage will be recursively merged with application configuration.

If one of the configuration parameters does not exist or can not be resolved, raises ArgumentError.

contains_jti?(server, jti)

contains_jti?(pid(), String.t()) :: boolean()

handle_call(arg, from, state)

handle_call({:contains?, String.t()}, any(), state_t()) ::
  {:reply, boolean(), state_t()}

handle_cast(arg, state)

handle_cast({:add, String.t(), Timex.DateTime.t(), nil | pid()}, state_t()) ::
  {:noreply, state_t()}

handle_info(arg1, state)

handle_info({:expire, String.t(), nil | pid()}, state_t()) ::
  {:noreply, state_t()}

handle_info(:expire_stale_records, state_t()) :: {:noreply, state_t()}

init(state)

init(state_t()) :: {:ok, state_t()}

start_link(tracker_mod \\ nil, opts \\ [])