Reactive Interaction Gateway v2.0.2 RigAuth.Session

Session tracking from authorization-header tokens.

Types

session_name_t()

session_name_t() :: String.t()

validity_period_t()

validity_period_t() :: pos_integer()

Functions

blacklist(session_name, validity_period_s)

blacklist(session_name_t(), validity_period_t()) :: nil

Disallow sessions with the given name for a specific amount of time.

blacklisted?(session_name)

blacklisted?(session_name_t()) :: boolean()

Check whether a session name has been disallowed.

config()

config() :: any()

Returns module configuration.

If application environment contains values in Keyword or Map struct, defaults from macro usage will be recursively merged with application configuration.

If one of the configuration parameters does not exist or can not be resolved, raises ArgumentError.

update(conn, pid)

update(Plug.Conn.t(), pid()) :: nil