Reactive Interaction Gateway v2.2.1 RigCloudEvents.CloudEvent

CloudEvents is a vendor-neutral specification for defining the format of event data.

See: https://github.com/cloudevents

Types

t()

t() :: %RigCloudEvents.CloudEvent{
  json: String.t(),
  parsed: RigCloudEvents.Parser.PartialParser.t()
}

Functions

find_value(cloud_event, json_pointer)

find_value(t(), json_pointer :: String.t()) ::
  {:ok, value :: any()} | {:error, any()}

id(event)

id!(event)

parse(json)

parse(String.t()) :: {:ok, t()} | {:error, any()}

parse(map()) :: {:ok, t()} | {:error, any()}

Convenience function used in testing.

If this would be called in production, of course it would be way more efficient to access the given map directly. However, this modules' raison d'être is the safe handling of incoming JSON encoded data, so it's safe to assume this function is ever only called by tests.

parse!(input)

parse!(String.t() | map()) :: t()

Initialize a new CloudEvent or raise.

See parse/1.

specversion(cloud_event)

specversion!(event)

type(event)

type!(event)