Reactive Interaction Gateway v2.2.1 RigInboundGateway.ApiProxy.Handler.Kafka

Handles requests for Kafka targets.

Functions

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

config()

config() :: any()

Returns module configuration.

If application environment contains values in Keyword or Map struct, defaults from macro usage will be recursively merged with application configuration.

If one of the configuration parameters does not exist or can not be resolved, raises ArgumentError.

do_handle_http_request(conn, request_path, partition, event, response_from)

handle_http_request(conn, api, endpoint, request_path)

Produce the request to a Kafka topic and optionally wait for the (correlated) response.

Expects a JSON encoded HTTP body with the following fields:

event: The published CloudEvent >= v0.2. The event is extended by metadata written to the "rig" extension field (following the CloudEvents v0.2 spec).

partition: The targetted Kafka partition.

kafka_handler(message)

kafka_handler(any()) ::
  :ok
  | {:error, %atom(){:__exception__ => true, optional(atom()) => any()},
     any()}

respond_with_bad_request(conn, response_from, description)

start_link(opts \\ [])

validate(conf)

validate(any()) :: {:ok, any()}