Reactive Interaction Gateway v2.4.0 RigInboundGateway.ApiProxy.Validations

Used to validate APIs configuration for reverse proxy. Validates use cases such as setting "target" to "kafka" but not "topic", which is required in such case.

When any error occurs during RIG start -> process will exit. When any error occurs during REST API request -> process won't exit, but instead API returns 400 -- bad request.

Types

error_list_t()

error_list_t() :: [{String.t(), error_t()}]

error_map_t()

error_map_t() :: %{
  required(String.t()) => [%{required(String.t() | atom()) => String.t()}]
}

error_t()

error_t() :: [{:error, String.t() | atom(), atom(), String.t()}]

Functions

config()

config() :: any()

Returns module configuration.

If application environment contains values in Keyword or Map struct, defaults from macro usage will be recursively merged with application configuration.

If one of the configuration parameters does not exist or can not be resolved, raises ArgumentError.

to_map(errors)

to_map(error_list_t()) :: error_map_t()

type_validation(map, key, type, fun)

type_validation(map(), String.t(), String.t(), function()) :: error_list_t()

validate(api)

validate(RigInboundGateway.ApiProxy.Api.t()) ::
  {:error, error_list_t()} | {:ok, RigInboundGateway.ApiProxy.Api.t()}

validate!(api)

validate!(RigInboundGateway.ApiProxy.Api.t()) ::
  RigInboundGateway.ApiProxy.Api.t()

validate_all(api)

validate_all(RigInboundGateway.ApiProxy.Api.t()) :: error_list_t()

validate_auth(errors, api)

validate_auth(error_list_t(), RigInboundGateway.ApiProxy.Api.t()) ::
  error_list_t()

validate_auth_type(api)

validate_auth_type(RigInboundGateway.ApiProxy.Api.t()) :: error_t()

validate_endpoint_target(endpoint, targets)

validate_endpoint_target(RigInboundGateway.ApiProxy.Api.endpoint(), [String.t()]) ::
  boolean()

validate_endpoints(errors, api)

validate_endpoints(error_list_t(), RigInboundGateway.ApiProxy.Api.t()) ::
  error_list_t()

validate_integer(map, key)

validate_integer(map(), String.t()) :: error_list_t()

validate_list(map, key)

validate_list(map(), String.t()) :: error_list_t()

validate_proxy(arg1)

validate_proxy(RigInboundGateway.ApiProxy.Api.t()) :: error_list_t()

validate_required_props(errors, api)

validate_required_props(error_list_t(), RigInboundGateway.ApiProxy.Api.t()) ::
  error_list_t()

validate_secured_endpoint(api, arg2)

validate_secured_endpoint(
  RigInboundGateway.ApiProxy.Api.t(),
  RigInboundGateway.ApiProxy.Api.endpoint()
) :: error_t()

validate_string(map, key)

validate_string(map(), String.t()) :: error_list_t()

validate_version_data(arg1)

validate_version_data(RigInboundGateway.ApiProxy.Api.t()) :: error_list_t()

with_any_error(errors, min_errors \\ 1)

with_any_error(error_list_t(), integer()) :: error_list_t()

with_nested_presence(arg1, key, map)

with_nested_presence(boolean(), String.t() | atom(), map()) :: error_t()