Reactive Interaction Gateway v2.0.2 RigInboundGateway.RateLimit

Allow only a certain amount of requests per seconds per endpoint (per IP).

For synchronizing the corresponding state between the short-lived request processes, an ETS table is used for optimal performance.

Functions

config()

config() :: any()

Returns module configuration.

If application environment contains values in Keyword or Map struct, defaults from macro usage will be recursively merged with application configuration.

If one of the configuration parameters does not exist or can not be resolved, raises ArgumentError.

request_passage(endpoint, ip \\ nil, opts \\ [])

request_passage(String.t(), String.t() | nil, %{} | []) :: :ok | :passage_denied

Request passage to a specific endpoint, from a given IP.

Depending on the granularity, the endpoint might be the hostname of the target host, or the target socket, e.g., hostname <> ":" <> port.

Calling this function always cause a side effect: internally, the request is recorded. Depending on previous calls, the function returns either :ok or :passage_denied. The latter means that the rate limit for the given endpoint (and source IP if per_ip? is true) has been reached, which means that the request in question should be blocked.