Reactive Interaction Gateway v2.0.0 RigInboundGateway.RateLimit.Sweeper

Periodically cleans up the ETS table.

By default, the remote IP is considered for rate-limiting, and, consequently, used within the ETS table key. This means that without cleanup, the table would grow quite large very fast.

The Sweeper cleans the table by removing all records that own a number of tokens equal to the configured burst size. This is okay because

removing records is atomic (per record)

if no record is found for a given endpoint and ip, it is (re-)created with the number of tokens equal to the burst size.

Can be disabled by setting the sweep interval to 0.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

start_link()

sweep(opts \\ [])

sweep([{:atom, any()}]) :: non_neg_integer()