RigInboundGatewayWeb.EventBuffer (Reactive Interaction Gateway v3.0.0-alpha.2)

A circular buffer for events.

Buffers events up to a configurable capacity. After that, every new item overwrites the oldest one.

Types

t()

Specs

t() :: %RigInboundGatewayWeb.EventBuffer{
  capacity: pos_integer(),
  events: [RigCloudEvents.CloudEvent.t()]
}

A circular event buffer.

Functions

add_event(event_buffer, event)

Specs

add_event(t(), RigCloudEvents.CloudEvent.t()) :: t()

Add an event to this buffer.

If the buffer runs at full capacity, this overwrites the oldest event in the buffer.

all_events(map)

Specs

all_events(t()) :: [RigCloudEvents.CloudEvent.t()]

All events, sorted from oldest to newest event.

capacity(map)

Specs

capacity(t()) :: pos_integer()

The capacity is the maximum number of events this buffer can hold.

events_since(map, event_id)

Specs

events_since(t(), event_id :: String.t()) ::
  {:ok, [events: [RigCloudEvents.CloudEvent.t()], last_event_id: String.t()]}
  | {:no_such_event, [not_found_id: String.t(), last_event_id: String.t()]}

new(capacity)

Specs

new(pos_integer()) :: t()

Creates a new EventBuffer with the given capacity.