Reactive Interaction Gateway v2.4.0 RigInboundGatewayWeb.V1.EventController

Publish a CloudEvent.

Functions

config()

config() :: any()

Returns module configuration.

If application environment contains values in Keyword or Map struct, defaults from macro usage will be recursively merged with application configuration.

If one of the configuration parameters does not exist or can not be resolved, raises ArgumentError.

publish(conn, params)

Plug action to send a CloudEvent to subscribers.