Reactive Interaction Gateway v2.0.2 RigInboundGatewayWeb.V1.Websocket

Cowboy WebSocket handler.

As soon as Phoenix pulls in Cowboy 2 this will have to be rewritten using the :cowboy_websocket behaviour.

Functions

websocket_handle(in_frame, state)

The client may send this as the response to the :ping heartbeat.