Reactive Interaction Gateway v2.2.1 RigOutboundGateway.Kinesis.JavaClient

Manages the external Java-based Kinesis client application.

In Java land this would've been named AmazonKinesisJavaClientManager.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

config()

config() :: any()

Returns module configuration.

If application environment contains values in Keyword or Map struct, defaults from macro usage will be recursively merged with application configuration.

If one of the configuration parameters does not exist or can not be resolved, raises ArgumentError.

handle_info(msg, state)

Starts (and awaits) the Java-client for Amazon Kinesis.

The process output is discarded. Instead of using stdout to receive Kinesis messages from the Java client, the Java code uses JInterface to RPC RigOutboundGateway.handle_raw/1 directly. This ensures that message boundaries are kept (think newlines in messages) and that console log output doesn't interfere.

java_client_callback(data)

java_client_callback(data :: [{atom(), String.t()}, ...]) :: :ok

start_link(opts)