Reactive Interaction Gateway v2.1.1 RIG.JWT behaviour

JSON Web Token handling.

Types

claims()

claims() :: %{optional(String.t()) => String.t()}

claims_and_errors()

claims_and_errors() :: [validation_result()]

ensure_not_blacklisted()

ensure_not_blacklisted() :: (claims() -> validation_result())

Turns claims into errors for blacklisted JWTs.

http_header()

http_header() :: {http_header_name :: String.t(), http_header_value()}

http_header_value()

http_header_value() :: String.t()

http_headers()

http_headers() :: [http_header()]

jwt_conf()

jwt_conf() :: %{alg: String.t(), key: String.t()}

token()

token() :: String.t()

validation_result()

validation_result() ::
  {:ok, claims()}
  | {:error, %RIG.JWT.DecodeError{__exception__: term(), cause: term()}}

Functions

config()

config() :: any()

Returns module configuration.

If application environment contains values in Keyword or Map struct, defaults from macro usage will be recursively merged with application configuration.

If one of the configuration parameters does not exist or can not be resolved, raises ArgumentError.

encode(claims, jwt_conf \\ config().jwt_conf())

encode(claims(), jwt_conf()) :: token()

parse_http_header(http_headers, jwt_conf \\ config().jwt_conf(), ensure_not_blacklisted \\ &ensure_not_blacklisted/1)

parse_token(token, jwt_conf \\ config().jwt_conf(), ensure_not_blacklisted \\ &ensure_not_blacklisted/1)

Callbacks

parse_http_header(arg1, jwt_conf, ensure_not_blacklisted)

parse_http_header(
  http_header_value() | http_headers(),
  jwt_conf(),
  ensure_not_blacklisted()
) :: claims_and_errors()

Find JWT claims in one or more HTTP headers.

All "Authorization" headers are considered. A single header may contain one or more credentials. Only "Bearer"-type (scheme) credentials are interpreted as JSON Web Tokens. Each of those JWTs is validated using their signature. The result contains JWT claims for successfully validated tokens and errors where the validation failed.

parse_token(token, jwt_conf, ensure_not_blacklisted)

parse_token(token(), jwt_conf(), ensure_not_blacklisted()) ::
  validation_result()

Extract claims from a given encoded JWT.