RIG.JWT behaviour (Reactive Interaction Gateway v3.0.0-alpha.2)

JSON Web Token handling.

Types

claims()

Specs

claims() :: %{optional(String.t()) => String.t()}

claims_and_errors()

Specs

claims_and_errors() :: [validation_result()]

ensure_not_blacklisted()

Specs

ensure_not_blacklisted() :: (claims() -> validation_result())

Turns claims into errors for blacklisted JWTs.

http_header()

Specs

http_header() :: {http_header_name :: String.t(), http_header_value()}

http_header_value()

Specs

http_header_value() :: String.t()

http_headers()

Specs

http_headers() :: [http_header()]

jwt_conf()

Specs

jwt_conf() :: %{alg: String.t(), key: String.t()}

token()

Specs

token() :: String.t()

validation_result()

Specs

validation_result() ::
  {:ok, claims()}
  | {:error, %RIG.JWT.DecodeError{__exception__: term(), cause: term()}}

Functions

config()

Specs

config() :: any()

Returns module configuration.

If application environment contains values in Keyword or Map struct, defaults from macro usage will be recursively merged with application configuration.

If one of the configuration parameters does not exist or can not be resolved, raises ArgumentError.

encode(claims, jwt_conf \\ config().jwt_conf)

Specs

encode(claims(), jwt_conf()) :: token()

See RIG.JWT.Claims.encode/2.

parse_http_header(http_headers, jwt_conf \\ config().jwt_conf, ensure_not_blacklisted \\ &ensure_not_blacklisted/1)

parse_token(token, jwt_conf \\ config().jwt_conf, ensure_not_blacklisted \\ &ensure_not_blacklisted/1)

valid?(token, jwt_conf \\ config().jwt_conf, ensure_not_blacklisted \\ &ensure_not_blacklisted/1)

Callbacks

parse_http_header(arg1, jwt_conf, ensure_not_blacklisted)

Specs

parse_http_header(
  http_header_value() | http_headers(),
  jwt_conf(),
  ensure_not_blacklisted()
) :: claims_and_errors()

Find JWT claims in one or more HTTP headers.

All "Authorization" headers are considered. A single header may contain one or more credentials. Only "Bearer"-type (scheme) credentials are interpreted as JSON Web Tokens. Each of those JWTs is validated using their signature. The result contains JWT claims for successfully validated tokens and errors where the validation failed.

parse_token(token, jwt_conf, ensure_not_blacklisted)

Specs

parse_token(token(), jwt_conf(), ensure_not_blacklisted()) ::
  validation_result()

Extract claims from a given encoded JWT.

valid?(token, jwt_conf, ensure_not_blacklisted)

Specs

valid?(token(), jwt_conf(), ensure_not_blacklisted()) :: boolean()

Checks an encoded JWT for validity.