Reactive Interaction Gateway v2.0.2 RigApi.Endpoint

Functions

broadcast(topic, event, msg)

Broadcasts a msg as event in the given topic.

Callback implementation for Phoenix.Endpoint.broadcast/3.

broadcast!(topic, event, msg)

Broadcasts a msg as event in the given topic.

Raises in case of failures.

Callback implementation for Phoenix.Endpoint.broadcast!/3.

broadcast_from(from, topic, event, msg)

Broadcasts a msg from the given from as event in the given topic.

Callback implementation for Phoenix.Endpoint.broadcast_from/4.

broadcast_from!(from, topic, event, msg)

Broadcasts a msg from the given from as event in the given topic.

Raises in case of failures.

Callback implementation for Phoenix.Endpoint.broadcast_from!/4.

call(conn, opts)

Callback implementation for Plug.call/2.

child_spec(opts)

Returns the child specification to start the endpoint under a supervision tree.

config(key, default \\ nil)

Returns the endpoint configuration for key

Returns default if the key does not exist.

config_change(changed, removed)

Reloads the configuration given the application environment changes.

host()

Returns the host for the given endpoint.

init(opts)

Callback implementation for Plug.init/1.

instrument(event, runtime \\ Macro.escape(%{}), fun) (macro)

Instruments the given function.

event is the event identifier (usually an atom) that specifies which instrumenting function to call in the instrumenter modules. runtime is metadata to be associated with the event at runtime (e.g., the query being issued if the event to instrument is a DB query).

Examples

instrument :render_view, %{view: "index.html"}, fn ->
  render(conn, "index.html")
end

path(path)

Generates the path information when routing to this endpoint.

script_name()

Generates the script name.

start_link(opts \\ [])

Starts the endpoint supervision tree.

static_path(path)

Generates a route to a static file in priv/static.

static_url()

Generates the static URL without any path information.

It uses the configuration under :static_url to generate such. It falls back to :url if :static_url is not set.

struct_url()

Generates the endpoint base URL but as a URI struct.

It uses the configuration under :url to generate such. Useful for manipulating the URL data and passing it to URL helpers.

subscribe(topic)

subscribe(pid, topic, opts)

unsubscribe(topic)

Unsubscribes the caller from the given topic.

Callback implementation for Phoenix.Endpoint.unsubscribe/1.

url()

Generates the endpoint base URL without any path information.

It uses the configuration under :url to generate such.