RigApi.Router.Helpers (Reactive Interaction Gateway v3.0.0-alpha.1)

Module with named helpers generated from RigApi.Router.

Functions

ap_is_path(conn_or_endpoint, action)

ap_is_path(conn_or_endpoint, action, params)

ap_is_path(conn_or_endpoint, action, id, params)

ap_is_url(conn_or_endpoint, action)

ap_is_url(conn_or_endpoint, action, params)

ap_is_url(conn_or_endpoint, action, id, params)

health_path(conn_or_endpoint, action)

health_path(conn_or_endpoint, action, params)

health_url(conn_or_endpoint, action)

health_url(conn_or_endpoint, action, params)

messages_path(conn_or_endpoint, action)

messages_path(conn_or_endpoint, action, params)

messages_url(conn_or_endpoint, action)

messages_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

responses_path(conn_or_endpoint, action)

responses_path(conn_or_endpoint, action, params)

responses_url(conn_or_endpoint, action)

responses_url(conn_or_endpoint, action, params)

session_blacklist_path(conn_or_endpoint, action)

session_blacklist_path(conn_or_endpoint, action, params)

session_blacklist_path(conn_or_endpoint, action, session_id, params)

session_blacklist_url(conn_or_endpoint, action)

session_blacklist_url(conn_or_endpoint, action, params)

session_blacklist_url(conn_or_endpoint, action, session_id, params)

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.