Reactive Interaction Gateway v2.4.0 RigInboundGateway.ApiProxy.PresenceHandler

Handles Phoenix Presence events.

Implemented as a Phoenix.Tracker, this module tracks all events it receives from the given Phoenix PubSub server. Only join events that refer to the @proxy topic are considered for inclusion into the PROXY.

The PROXY server is started from, and linked against, this server.

Functions

handle_diff(diff, state)

Forwards joins on the "proxy" topic to the Proxy server.

Leaves are not handled since we want to keep track of removed APIs and potentially renew them.

start_link(opts)