Reactive Interaction Gateway v2.3.0 RigInboundGatewayWeb.EventBuffer

A circular buffer for events.

Buffers events up to a configurable capacity. After that, every new item overwrites the oldest one.

Types

t()

t() :: %RigInboundGatewayWeb.EventBuffer{
  capacity: pos_integer(),
  events: [RigCloudEvents.CloudEvent.t()]
}

A circular event buffer.

Functions

add_event(event_buffer, event)

add_event(RigInboundGatewayWeb.EventBuffer.t(), RigCloudEvents.CloudEvent.t()) ::
  RigInboundGatewayWeb.EventBuffer.t()

Add an event to this buffer.

If the buffer runs at full capacity, this overwrites the oldest event in the buffer.

all_events(map)

all_events(RigInboundGatewayWeb.EventBuffer.t()) :: [
  RigCloudEvents.CloudEvent.t()
]

All events, sorted from oldest to newest event.

capacity(map)

capacity(RigInboundGatewayWeb.EventBuffer.t()) :: pos_integer()

The capacity is the maximum number of events this buffer can hold.

events_since(map, event_id)

events_since(RigInboundGatewayWeb.EventBuffer.t(), event_id :: String.t()) ::
  {:ok, [events: [RigCloudEvents.CloudEvent.t()], last_event_id: String.t()]}
  | {:no_such_event, [not_found_id: String.t(), last_event_id: String.t()]}

new(capacity)

new(pos_integer()) :: RigInboundGatewayWeb.EventBuffer.t()

Creates a new EventBuffer with the given capacity.