Reactive Interaction Gateway v3.0.0-alpha.1 RigInboundGatewayWeb.Router.Helpers

Module with named helpers generated from RigInboundGatewayWeb.Router.

Functions

event_path(conn_or_endpoint, action)

event_path(conn_or_endpoint, action, params)

event_url(conn_or_endpoint, action)

event_url(conn_or_endpoint, action, params)

health_path(conn_or_endpoint, action)

health_path(conn_or_endpoint, action, params)

health_url(conn_or_endpoint, action)

health_url(conn_or_endpoint, action, params)

longpolling_path(conn_or_endpoint, action)

longpolling_path(conn_or_endpoint, action, params)

longpolling_url(conn_or_endpoint, action)

longpolling_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

subscription_path(conn_or_endpoint, action, connection_id)

subscription_path(conn_or_endpoint, action, connection_id, params)

subscription_url(conn_or_endpoint, action, connection_id)

subscription_url(conn_or_endpoint, action, connection_id, params)

url(data)

Generates the connection/endpoint base URL without any path information.